+++ src/gui.js
@@ -87,11 +87,11 @@ HatBlockMorph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2025-January-24';
+modules.gui = '2025-January-25';
 
 // Declarations
 
-var SnapVersion = '10.4.4';
+var SnapVersion = '10.4.5';
 
 var IDE_Morph;
 var ProjectDialogMorph;