+++ sw.js
@@ -1,6 +1,6 @@
 /*global self, caches*/
 /*jshint esversion: 6*/
-var snapVersion = '10.4.4',
+var snapVersion = '10.4.5',
     cacheName = `snap-pwa-${snapVersion}`,
     filesToCache = [
         'snap.html',